requirements.txt
fastapi==0.103.2
pillow==10.4.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-multipart==0.0.6